test(refactor): move to concern unit and module rspec
